worldmodels1/pyproject.toml
[tool.poetry]
name = "worldmodels1"
version = "0.1.0"
description = ""
authors = ["Dean <deanvanaswegen.dva@gmail.com>"]
readme = "README.md"
packages = [{include = "worldmodels1", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
swig = "^4.1.1"
gym = "^0.26.2"
ipykernel = "^6.25.2"
matplotlib = "^3.8.0"
pillow = "^10.0.1"
numpy = "<=1.26.0"
pandas = "^2.1.0"
seaborn = "^0.12.2"
torch = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
